+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # CHANGELOG
 
 
+## v0.1.10
+* Add `pbjUrl` which uses URI.js (http://medialize.github.io/URI.js/) to provide url resolution for pbj instances using templates.
+
+
 ## v0.1.9
 * Adjust `PbjxToken` TTL to 10 (down from 120) and LEEWAY to 300 (from 30).  Because LEEWAY is used in both iat and exp validation we need that window to be larger, exp works within that expanded window as well.